backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic-settings==2.8.1
supabase==2.15.0
anthropic==0.50.0
httpx==0.28.1
reportlab==4.4.0
